workerNode/requirements.txt
click==6.7
Flask==1.1.2
numpy==1.20.1
Dlib==19.21.1
opencv-python==4.5.1.48
itsdangerous==0.24
Jinja2>=2.10.1
MarkupSafe==1.0
Werkzeug==2.0.3
imutils==0.5.3
pyzmq>=16
python-dotenv==1.0.0